# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jipe Life MLM Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/validate-sponsor:
    get:
      operationId: validateSponsor
      tags: [auth]
      summary: Validate sponsor ID and return sponsor name
      parameters:
        - name: sponsorId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sponsor info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorInfo"
        "404":
          description: Sponsor not found
  /auth/validate-epin:
    get:
      operationId: validateEpin
      tags: [auth]
      summary: Validate E-Pin and return package info
      parameters:
        - name: pin
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: E-Pin package info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpinValidation"
        "400":
          description: Invalid or used pin
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user with E-Pin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with member ID and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Admin login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /epins:
    get:
      operationId: listMyEpins
      tags: [epins]
      summary: List current user's E-Pins
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [unused, used, transferred]
      responses:
        "200":
          description: E-Pin list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Epin"
  /epins/transfer:
    post:
      operationId: transferEpin
      tags: [epins]
      summary: Transfer an E-Pin to a downline member (OTP verified)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpinTransferInput"
      responses:
        "200":
          description: Transfer successful
        "400":
          description: Transfer failed
...
  /rewards/claim:
    post:
      operationId: claimReward
      tags: [rewards]
      summary: Claim a reward at a milestone (choose reward or cash)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardClaimInput"
      responses:
        "200":
          description: Claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardClaim"

  # ── ADMIN ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SponsorInfo:
      type: object
      required: [memberId, name]
      properties:
        memberId:
          type: string
        name:
          type: string
        package:
          type: string
    EpinValidation:
      type: object
      required: [pin, package, packagePrice, bv, valid]
      properties:
        pin:
          type: string
        package:
          type: string
          enum: [smart, silver, gold]
        packagePrice:
          type: integer
        bv:
          type: integer
        valid:
          type: boolean
    RegisterInput:
      type: object
      required: [sponsorId, epin, name, email, phone, password]
      properties:
        sponsorId:
          type: string
        epin:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        position:
          type: string
          enum: [left, right]
    LoginInput:
      type: object
      required: [memberId, password]
      properties:
        memberId:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, memberId, name, email, phone, package, bv, status, role, createdAt]
      properties:
        id:
          type: integer
        memberId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        package:
          type: string
          enum: [smart, silver, gold]
        bv:
          type: integer
        status:
          type: string
          enum: [active, blocked, inactive]
        role:
          type: string
          enum: [member, admin]
        sponsorId:
          type: ["integer", "null"]
        leftChildId:
          type: ["integer", "null"]
        rightChildId:
          type: ["integer", "null"]
        ctoActive:
          type: boolean
        createdAt:
          type: string
...